+++ packages/cursorless-engine/src/disabledComponents/DisabledTalonSpokenForms.ts
@@ -1,0 +1,11 @@
+import type { TalonSpokenForms } from "../scopeProviders/TalonSpokenForms";
+
+export class DisabledTalonSpokenForms implements TalonSpokenForms {
+  getSpokenFormEntries() {
+    return Promise.resolve([]);
+  }
+
+  onDidChange() {
+    return { dispose: () => {} };
+  }
+}
